refactor: migrate llm providers to ai.common.LLMBase
