+++ nodes/src/nodes/llm_openai_api/IInstance.py
@@ -1,5 +1,5 @@
-from nodes.llm_base import IInstanceGenericLLM
+from ai.common.llm_base import LLMBase
 
 
-class IInstance(IInstanceGenericLLM):
+class IInstance(LLMBase):
     pass
