+++ nodes/src/nodes/llm_qwen/IInstance.py
@@ -21,8 +21,8 @@
 # SOFTWARE.
 # =============================================================================
 
-from nodes.llm_base import IInstanceGenericLLM
+from ai.common.llm_base import LLMBase
 
 
-class IInstance(IInstanceGenericLLM):
+class IInstance(LLMBase):
     pass
