+++ nodes/src/nodes/llm_vision_ollama/IInstance.py
@@ -22,12 +22,12 @@
 # =============================================================================
 
 from .IGlobal import IGlobal
-from nodes.llm_base import IInstanceGenericLLM
+from ai.common.llm_base import LLMBase
 from rocketlib import AVI_ACTION, warning
 from ai.common.schema import Doc
 
 
-class IInstance(IInstanceGenericLLM):
+class IInstance(LLMBase):
     """Instance handler for the Ollama Vision node."""
 
     IGlobal: IGlobal
